--- a/pkg/col/colserde/file.go
+++ b/pkg/col/colserde/file.go
@@ -72,6 +72,10 @@ func NewFileSerializer(w io.Writer, typs []coltypes.T) (*FileSerializer, error) 
 	return s, s.Reset(w)
 }
 
+func (s *FileSerializer) Written() int {
+	return s.w.written
+}
+
 // Reset can be called to reuse this FileSerializer with a new io.Writer after
 // calling Finish. The types will remain the ones passed to the constructor. The
 // caller is responsible for closing the given writer.
@@ -124,7 +128,8 @@ func (s *FileSerializer) AppendBatch(batch coldata.Batch) error {
 }
 
 // Finish writes the footer metadata described by the arrow spec. Nothing can be
-// called after Finish except Reset.
+// called after Finish except Reset. The total number of bytes written since
+// the serializer was initialized is returned.
 func (s *FileSerializer) Finish() error {
 	defer func() {
 		s.w = nil