Release note: None

Here is the branch with the benchmarks. To compare both the pebble and file implementations I use:
```
for i in 4KiB 8KiB 16KiB 32KiB 64KiB 128KiB 256KiB 512KiB 1MiB 2MiB 4MiB
do
    echo "running $i"
    pfname=cpeb$i
    ffname=cfl$i
    make bench PKG=./pkg/col/colserde BENCHES=BenchmarkQueue TESTFLAGS="-v -count 10 -store=pebble -bufsize=$i -blocksize=512KiB -datasize=512MiB" > $pfname
    make bench PKG=./pkg/col/colserde BENCHES=BenchmarkQueue TESTFLAGS="-v -count 10 -store=flat -bufsize=$i -blocksize=64MiB -datasize=512MiB" > $ffname
    benchstat $pfname $ffname
done
```