@asubiotto 2019-11-21
With the addition of `NewSyncingFile` in the `DiskQueue` prototype (in `rotateFile`), the file-based implementation seems to perform much worse than pebble. This was run on a `gceworker`:
```
running 4KiB
name       old time/op  new time/op  delta
Queues-24   2.10s ± 7%  11.14s ± 0%  +431.11%  (p=0.000 n=9+7)
running 8KiB
name       old time/op  new time/op  delta
Queues-24   2.07s ± 2%  11.13s ± 0%  +436.67%  (p=0.000 n=8+7)
running 16KiB
name       old time/op  new time/op  delta
Queues-24   2.10s ± 7%  10.97s ± 0%  +422.03%  (p=0.000 n=8+7)
running 32KiB
name       old time/op  new time/op  delta
Queues-24   2.10s ± 9%  10.92s ± 0%  +420.14%  (p=0.000 n=8+8)
running 64KiB
name       old time/op  new time/op  delta
Queues-24   2.10s ± 7%  10.87s ± 0%  +418.24%  (p=0.000 n=8+7)
...
```
This is kind of reminiscent of how long it took to store and re-read `512MiB` with a `32MiB` in-memory write buffer on top of pebble https://github.com/cockroachdb/pebble/issues/404 although it could be completely orthogonal.