@petermattis 2019-11-21
> This is kind of reminiscent of how long it took to store and re-read 512MiB with a 32MiB in-memory write buffer on top of pebble cockroachdb/pebble#404 although it could be completely orthogonal.

That problem is due to an implementation detail of `Pebble` memtables. Whatever is going on here is different.

I took a look at this PR and don't see anything obviously wrong. And I verified that disabling syncing provides a significant speedup. I bet that delta would drop if we increased `--datasize` so that it is larger than the memory size of the machine. Or wrote a program to `mlock` a large chunk of memory in order to simulate a smaller amount of RAM.