@asubiotto 2019-11-22
I haven't tried what you suggested yet (traveling with spotty internet), but I extended `syncingFile` to record the number of times `Sync` is actually called as well as some additional statistics about whether we returned early in `syncingFile.maybeSync` (on `macOS` with `512KiB` buffer size and `512MiB` data size):

| | Pebble | File |
|-----|-----|-----|
| Total times in `maybeSync` | 2021 | 2040 |
| Times actually calling sync | 88 (~4%) | 336 (~16%) |
| Not syncing because too close to write offset | 57 (~3%) | 16 (~1%) |
| Not syncing due to bytes below threshold | 1882 (~93%) | 1688 (~83%) |
| Total bytes written through `Write` | `128MiB` | `512MiB` |
| Total bytes synced? (cum(syncToOffset-syncOffset)) | `~46mB` | `~178mB` |

Based on the benchmark numbers and number of syncs, it seems like the average cost per `Sync` is equal in both cases (which is expected), the problem is that the file-based implementation is `Sync`ing a lot more often, and it seems that it's due to the `Pebble` implementation `Write`ing less bytes. This seemed to be due to compression, so I disabled it:
```
	opts.Levels = []pebble.LevelOptions{
		{
			Compression: pebble.NoCompression,
		},
	}
```
And it so far seems like the benchmarks have improved (on `macOS`):
```
running 64KiB
name       old time/op  new time/op  delta
Queues-16   5.99s ± 1%   5.29s ± 4%  -11.67%  (p=0.000 n=9+10)
running 128KiB
name       old time/op  new time/op  delta
Queues-16   6.04s ± 5%   5.04s ± 4%  -16.53%  (p=0.000 n=9+10)
running 256KiB
name       old time/op  new time/op  delta
Queues-16   5.99s ± 2%   4.97s ± 6%  -17.17%  (p=0.000 n=10+9)
```
Will benchmark further etc... but it seems like we might want to add compression to the file-based queue.