@petermattis 2019-11-22
> Will benchmark further etc... but it seems like we might want to add compression to the file-based queue.

Good find. Adding snappy compression is pretty easy. See `github.com/golang/snappy`.